requirements.txt
allure-pytest>=2.13.5
pytest>=9.0.0
python-dotenv>=1.0.0
selenium>=4.0.0
